Split dependency determination out into common
